Renamed module io to metadata
